Ihms/requirements.txt
asgiref==3.8.1
awscli==1.35.6
boto3==1.35.40
botocore==1.35.40
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docutils==0.16
ecdsa==0.19.0
Faker==30.3.0
idna==3.10
jmespath==1.0.1
MarkupSafe==3.0.1
pyasn1==0.6.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.2
requests==2.32.3
rsa==4.7.2
s3transfer==0.10.3
serverless-wsgi==3.0.4
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
